% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{AverageAndRegularize}
\alias{AverageAndRegularize}
\title{Perform comparative analyses of kinetic parameter estimates}
\usage{
AverageAndRegularize(
  obj,
  features = "all",
  parameter = "log_kdeg",
  formula_mean = NULL,
  formula_sd = NULL,
  include_all_parameters = TRUE,
  sd_reg_factor = 10,
  error_if_singular = TRUE
)
}
\description{
Perform comparative analyses of kinetic parameter estimates
}
