% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dropout.R
\name{CorrectDropout}
\alias{CorrectDropout}
\title{Correct for experimental/bioinformatic dropout of labeled RNA.}
\usage{
CorrectDropout(
  obj,
  grouping_factors = NULL,
  features = NULL,
  exactMatch = TRUE
)
}
\arguments{
\item{obj}{An EZbakRFractions object, which is an EZbakRData object on which
you have run \code{EstimateFractions()}.}

\item{grouping_factors}{Which sample-detail columns in the metadf should be used
to group -s4U samples by for calculating the average -s4U RPM? The default value of
\code{NULL} will cause all sample-detail columns to be used.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of \code{NULL}
will expect there to be only one fractions table in the EZbakRFractions object.}

\item{exactMatch}{If TRUE, then \code{features} must exactly match the \code{features}
metadata for a given fractions table for it to be used. Means that you cannot
specify a subset of features by default. Set this to FALSE if you would like
to specify a feature subset.}
}
\description{
Correct for experimental/bioinformatic dropout of labeled RNA.
}
