% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateFractions.EZbakRData}
\alias{EstimateFractions.EZbakRData}
\title{Estimate fractions of each RNA population with standard EZbakRData object}
\usage{
\method{EstimateFractions}{EZbakRData}(
  obj,
  features = "all",
  mutrate_populations = "all",
  fraction_design = NULL,
  Poisson = TRUE,
  strategy = c("standard"),
  column_to_reorder = NULL,
  pnew_prior_mean = -2.94,
  pnew_prior_sd = 0.3,
  pold_prior_mean = -6.5,
  pold_prior_sd = 0.5
)
}
\description{
Estimate fractions of each RNA population with standard EZbakRData object
}
