% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinetics.R
\name{EstimateKinetics}
\alias{EstimateKinetics}
\title{Generic function for estimating kinetic parameters}
\usage{
EstimateKinetics(obj, features = "all", strategy = "standard")
}
\description{
Generic function for estimating kinetic parameters
}
