% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateMutRates.EZbakRArrowData}
\alias{EstimateMutRates.EZbakRArrowData}
\title{Estimate mutation rates}
\usage{
\method{EstimateMutRates}{EZbakRArrowData}(
  obj,
  populations = "all",
  pnew_prior_mean = -2.94,
  pnew_prior_sd = 0.3,
  pold_prior_mean = -6.5,
  pold_prior_sd = 0.5,
  pold_est = NULL
)
}
\arguments{
\item{obj}{An \code{EZbakRData} object}

\item{populations}{Character vector of the set of mutational populations
that you want to infer the fractions of. For example, say your cB file contains
columns tracking T-to-C and G-to-A}

\item{pnew_prior_mean}{logit-Normal mean for logit(pnew) prior.}

\item{pnew_prior_sd}{logit-Normal sd for logit(pnew) prior.}

\item{pold_prior_mean}{logit-Normal mean for logit(pold) prior.}

\item{pold_prior_sd}{logit-Normal sd for logit(pold) prior.}

\item{pold_est}{Background mutation rate estimates if you have them. Can either be a single
number applied to all samples or a named vector of values, where the names should be sample
names.}
}
\description{
Two component mixture models are fit to all data to estimate global high and
low mutation rates for all samples. Estimation of these mutation rates
are regularized through the use of informative priors whose parameters can
be altered using the arguments defined
}
\details{
This method expects the input object to contain an on-disk Arrow Dataset in place
of an in-memory cB file. Furtheromore, this dataset is expected to be partitioned by
the individual samples. This allows EZbakR to load only a single sample worth of data
in to memory at a time, which can significantly reduce RAM usage. The input object
should be created with \code{EZbakRArrowData()}.
}
