% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fraction_New.R
\name{EstimateMutRates}
\alias{EstimateMutRates}
\title{Estimate mutation rates}
\usage{
EstimateMutRates(
  obj,
  populations = "all",
  pnew_strategy = "standard",
  pold_strategy = "standard"
)
}
\arguments{
\item{obj}{EZbakRDataobject}

\item{populations}{Character vector of the set of mutational populations
that you want to infer the fractions of. For example, say your cB file contains
columns tracking T-to-C and G-to-A}

\item{pnew_strategy}{String denoting which new read mutation rate estimation strategy to use.
Options include:
\itemize{
\item standard: Estimate a single new read and old read mutation rate for each
sample. This is done via a binomial mixture model aggregating over
\item intronic: Estimate sample-wide pnew from intronic mutation rate. Have to
specify which feature in the EZbakRDataobject cB represents the intronic reads
\item hierarchical: Estimate feature-specific pnew with standard , regularizing the feature-specific
estimate with a sample-wide prior.
}}

\item{pold_strategy}{String denoting which old read mutation rate estimation strategy}
}
\description{
Estimate mutation rates
}
