% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_cB}
\alias{example_cB}
\title{Example cB table}
\format{
\subsection{\code{example_cB}}{

A tibble with 10,000 rows and 7 columns:
\describe{
\item{sample}{Sample name}
\item{rname}{Chromosome name}
\item{GF}{Gene name for reads aligning to any region of a gene}
\item{XF}{Gene name for reads aligning to exclusively exonic regions of a gene}
\item{TC}{Number of T-to-C mutations}
\item{nT}{Number of Ts}
\item{n}{Number of reads with same value for the first 6 columns}
}
}
}
\usage{
example_cB
}
\description{
An example \code{cB} table used to create an \code{EZbakRData} object. This cB table is a
subset of a cB file from the DCP2 KO dataset published in Luo et al., 2020.
The original file is large (69 MB), so the example cB file has been
downsampled and contains only a subset of reads from chromosome 21.
}
\references{
Luo et al. (2020) Biochemistry. 59(42), 4121-4142
}
\keyword{datasets}
