% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_metadf}
\alias{example_metadf}
\title{Example metadf}
\format{
\subsection{\code{example_metadf}}{

A tibble with 6 rows and 3 columns
\describe{
\item{sample}{Sample name}
\item{tl}{Metabolic label feed time}
\item{genotype}{Whether sample was collected from WT or DCP2 KO cells}
}
}
}
\usage{
example_metadf
}
\description{
An example \code{metadf} table used to create an \code{EZbakRData} object. This metadf
describes the DCP2 KO dataset published in Luo et al., 2020.
}
\references{
Luo et al. (2020) Biochemistry. 59(42), 4121-4142
}
\keyword{datasets}
