% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{EZMAPlot}
\alias{EZMAPlot}
\title{Make an MAPlot from EZbakR comparison}
\usage{
EZMAPlot(
  obj,
  parameter = "log_kdeg",
  design_factor = NULL,
  reference = NULL,
  experimental = NULL,
  param_name = NULL,
  param_function = NULL,
  features = NULL,
  condition = NULL,
  exactMatch = TRUE,
  plotlog2 = TRUE,
  FDR_cutoff = 0.05,
  difference_cutoff = 0,
  size = NULL
)
}
\arguments{
\item{obj}{An object of class \code{EZbakRCompare}, which is an \code{EZbakRData} object
on which you have run \code{CompareParameters}}

\item{parameter}{Name of parameter whose comparison you want to plot.}

\item{design_factor}{Name of factor from \code{metadf} whose parameter estimates at
different factor values you would like to compare.}

\item{reference}{Name of reference \code{condition} factor level value.}

\item{experimental}{Name of \code{condition} factor level value to compare to reference.}

\item{param_name}{If you want to assess the significance of a single parameter,
rather than the comparison of two parameters, specify that one parameter's name
here.}

\item{param_function}{NOT YET IMPLEMENTED. Will allow you to specify more complicated
functions of parameters when hypotheses you need to test are combinations of parameters
rather than individual parameters or simple differences in two parameters.}

\item{features}{Character vector of feature names for which comparisons were made.}

\item{condition}{Defunct parameter that has been replaced with \code{design_factor}. If provided
gets passed to \code{design_factor} if \code{design_factor} is not already specified.}

\item{exactMatch}{If TRUE, then \code{features} and \code{populations} have to exactly match
those for a given fractions table for that table to be used. Means that you can't
specify a subset of features or populations by default, since this is TRUE
by default.}

\item{plotlog2}{If TRUE, assume that log(parameter) difference is passed in and that
you want to plot log2(parameter) difference.}

\item{FDR_cutoff}{False discovery cutoff by which to color points.}

\item{difference_cutoff}{Minimum absolute difference cutoff by which to color points.}

\item{size}{Size of points, passed to \code{geom_point()} size parameter. If not specified,
a point size is automatically chosen.}
}
\description{
Make an MAPlot from EZbakR comparison
}
