% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{EZQC.EZbakRData}
\alias{EZQC.EZbakRData}
\title{Run quality control checks}
\usage{
\method{EZQC}{EZbakRData}(obj, mutrate_populations = "all", ..., features = "all")
}
\description{
Run quality control checks
}
