% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{EZQC.EZbakRFractions}
\alias{EZQC.EZbakRFractions}
\title{Run quality control checks}
\usage{
\method{EZQC}{EZbakRFractions}(
  obj,
  mutrate_populations = "all",
  ...,
  features = NULL,
  populations = NULL,
  fraction_design = NULL
)
}
\description{
Run quality control checks
}
