% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{EZVolcanoPlot}
\alias{EZVolcanoPlot}
\title{Make a VolcanoPlot from EZbakR comparison}
\usage{
EZVolcanoPlot(
  obj,
  parameter = "log_kdeg",
  design_factor = NULL,
  reference = NULL,
  experimental = NULL,
  param_name = NULL,
  param_function = NULL,
  features = NULL,
  condition = NULL,
  normalize_by_median = NULL,
  repeatID = NULL,
  exactMatch = TRUE,
  plotlog2 = TRUE,
  FDR_cutoff = 0.05,
  difference_cutoff = log(2),
  size = NULL
)
}
\arguments{
\item{obj}{An object of class \code{EZbakRCompare}, which is an \code{EZbakRData} object
on which you have run \code{CompareParameters}}

\item{parameter}{Name of parameter whose comparison you want to plot.}

\item{design_factor}{Name of factor from \code{metadf} whose parameter estimates at
different factor values you would like to compare.}

\item{reference}{Name of reference \code{condition} factor level value.}

\item{experimental}{Name of \code{condition} factor level value to compare to reference.}

\item{param_name}{If you want to assess the significance of a single parameter,
rather than the comparison of two parameters, specify that one parameter's name
here.}

\item{param_function}{NOT YET IMPLEMENTED. Will allow you to specify more complicated
functions of parameters when hypotheses you need to test are combinations of parameters
rather than individual parameters or simple differences in two parameters.}

\item{features}{Character vector of feature names for which comparisons were made.}

\item{condition}{Defunct parameter that has been replaced with \code{design_factor}. If provided
gets passed to \code{design_factor} if \code{design_factor} is not already specified.}

\item{normalize_by_median}{Whether or not the median was subtracted from the estimated
parameter differences.}

\item{repeatID}{If multiple \code{kinetics} or \code{fractions} tables exist with the same metadata,
then this is the numerical index by which they are distinguished.}

\item{exactMatch}{If TRUE, then \code{features} has to exactly match
those for a given comparisons table for that table to be used. Means that you can't
specify a subset of features by default, since this is TRUE
by default.}

\item{plotlog2}{If TRUE, assume that log(parameter) difference is passed in and that
you want to plot log2(parameter) difference. TO-DO: probably best to change this
to a more general scale parameter by which the parameter is multiplied. Default
would be log2(exp(1)) to convert log() to log2().}

\item{FDR_cutoff}{False discovery cutoff by which to color points.}

\item{difference_cutoff}{Minimum absolute difference cutoff by which to color points.}

\item{size}{Size of points, passed to \code{geom_point()} size parameter. If not specified,
a point size is automatically chosen.}
}
\value{
A \code{ggplot2} object. X-axis = log2(estimate of interest (e.g., fold-change
in degradation rate constant); Y-axis = -log10(multiple test adjusted p-value);
points colored by location relative to FDR and effect size cutoffs.
}
\description{
Make a plot of effect size (x-axis) vs. multiple-test adjusted p-value (y-axis),
coloring points by position relative to user-defined decision cutoffs.
}
\details{
\code{EZVolcanoPlot()} accepts as input the output of \code{CompareParameters()}, i.e.,
an \code{EZbakRData} object with at least one "comparisons" table. It will plot
the "difference" column in this table versus -log10 of the "padj" column.
In the simplest case, "difference" represents a log-fold change in a kinetic
parameter (e.g., kdeg) estimate. More complicated linear model fits and
comparisons can yield different parameter estimates.

NOTE: some outputs of \code{CompareParameters()} are not meant for visualization
via a volcano plot. For example, when fitting certain interaction models,
some of the parameter estimates may represent average log(kinetic paramter)
in one condition. See discussion of one example of this \href{https://github.com/simonlabcode/bakR/issues/7#issuecomment-2371431127}{here}.

EZbakR estimates kinetic parameters in \code{EstimateKinetics()} and \code{EZDynamics()}
on a log-scale. By default, since log2-fold changes are a bit easier to interpret
and more common for these kind of visualizations, \code{EZVolcanoPlot()} multiplies
the x-axis value by log2(exp(1)), which is the factor required to convert from
a log to a log2 scale. You can turn this off by setting \code{plotlog2} to \code{FALSE}.
}
