% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isoforms.R
\name{ImportIsoformQuant}
\alias{ImportIsoformQuant}
\title{Import transcript isoform quantification into EZbakRData object}
\usage{
ImportIsoformQuant(
  obj,
  files,
  quant_tool = c("none", "salmon", "sailfish", "alevin", "piscem", "kallisto", "rsem",
    "stringtie"),
  ...
)
}
\description{
A convenient wrapper to \code{tximport()} for importing isoform quantification
data into an EZbakRData object. This is necessary to run functions such as
\code{EstimateIsoformFractions}.
}
