% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRArrowData.R
\name{new_EZbakRArrowData}
\alias{new_EZbakRArrowData}
\title{\code{EZbakRArrowData} object constructor for internal use}
\usage{
new_EZbakRArrowData(cBds, metadf)
}
\arguments{
\item{cBds}{Arrow dataset tracking the sample ID, mutational and nucleotide content,
and feature assignment of sequencing reads.}

\item{metadf}{Data frame tracking features of each of the samples included
in \code{cBds}.}
}
\description{
\code{new_EZbakRArrowData} efficiently creates an object of class \code{EZbakRArrowData}.
It does not perform any rigorous checks of the legitimacy of this object.
}
