% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRData.R
\name{new_EZbakRData}
\alias{new_EZbakRData}
\title{\code{EZbakRDataobject} constructor for internal use}
\usage{
new_EZbakRData(cB, metadf)
}
\arguments{
\item{cB}{Data frame tracking the sample ID, mutational and nucleotide content,
and feature assignment of sequencing reads.}

\item{metadf}{Data frame tracking features of each of the samples included
in \code{cB}.}
}
\description{
\code{new_EZbakRData} efficiently creates an object of class \code{EZbakRData}.
It does not perform any rigorous checks of the legitimacy of this object.
}
