% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRFractions.R
\name{new_EZbakRFractions}
\alias{new_EZbakRFractions}
\title{\code{EZbakRFractions} object constructor}
\usage{
new_EZbakRFractions(fractions, metadf, name = NULL, character_limit = 20)
}
\arguments{
\item{fractions}{Data frame containing information about the fraction of reads
from each mutational population of interest.}

\item{metadf}{Data frame reporting aspects of each of the samples included
in \code{fractions}}
}
\description{
\code{new_EZbakRFractions} efficiently creates an object of class \code{EZbakRFractions}.
It does not perform any rigorous checks of the legitimacy of this object.
}
