% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRKinetics.R
\name{new_EZbakRKinetics}
\alias{new_EZbakRKinetics}
\title{\code{EZbakRKinetics} object constructor}
\usage{
new_EZbakRKinetics(
  kinetics,
  features,
  metadf,
  name = NULL,
  character_limit = 20
)
}
\arguments{
\item{kinetics}{Data frame containing information about the kinetic parameters
of interest for each set of features tracked.}

\item{features}{Features tracked in \code{kinetics} data frame. Needs to be specified
explicitly as it cannot be automatically inferred.}

\item{metadf}{Data frame describing each of the samples included}

\item{name}{Optional; name to give to fractions table.}

\item{character_limit}{Maximum number of characters for naming out fractions output. EZbakR
will try to name this as a "_" separated character vector of all of the features analyzed.
If this name is greater than \code{character_limit}, then it will default to "fraction#", where
"#" represents a simple numerical ID for the table.
in \code{kinetics}}
}
\description{
\code{new_EZbakRData} efficiently creates an object of class \code{EZbakRData}.
It does not perform any rigorous checks of the legitimacy of this object.
}
