% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{print.EZbakRArrowData}
\alias{print.EZbakRArrowData}
\title{Print method for \code{EZbakRArrowData} objects}
\usage{
\method{print}{EZbakRArrowData}(x, max_name_chars = 60, ...)
}
\arguments{
\item{x}{An \code{EZbakRArrowData} object.}

\item{max_name_chars}{Maximum number of characters to print on each line}

\item{...}{Ignored}
}
\value{
The input \code{EZbakRData} object, invisibly
}
\description{
Print method for \code{EZbakRArrowData} objects
}
\examples{

# Simulate data to analyze
simdata <- SimulateOneRep(30)

# Create EZbakR input
metadf <- data.frame(sample = "sampleA", tl = 2)
ezbdo <- EZbakRData(simdata$cB, metadf)

# Print
print(ezbdo)

}
