% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{print.EZbakRData}
\alias{print.EZbakRData}
\title{Print method for \code{EZbakRData} objects}
\usage{
\method{print}{EZbakRData}(x, max_name_chars = 60, ...)
}
\arguments{
\item{x}{An \code{EZbakRData} object.}

\item{max_name_chars}{Maximum number of characters to print on each line}

\item{...}{Ignored}
}
\description{
Print method for \code{EZbakRData} objects
}
