% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple-functions.R
\name{SimpleSim}
\alias{SimpleSim}
\title{Simple simulation function}
\usage{
SimpleSim(
  nreads = 1000,
  fn = 0.5,
  pnew = 0.05,
  pold = 0.001,
  rlen = 100,
  Ucont = 0.25
)
}
\arguments{
\item{nreads}{Number of reads to simulate}

\item{fn}{Fraction of reasd that are new in simulation. Whether a read will
be new will be determined by a draw from a Bernoulli(fn) distribution.}

\item{pnew}{T-to-C mutation rate in new reads}

\item{pold}{T-to-C mutation rate in old reads}

\item{rlen}{Length of simulated reads}

\item{Ucont}{Fraction of nucleotides in simulated reads that are Ts (U in RNA)}
}
\description{
Simple simulation function
}
