% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{SimulateDynamics}
\alias{SimulateDynamics}
\title{Simulation of generalized dynamical system model.}
\usage{
SimulateDynamics(
  nfeatures,
  graph,
  metadf,
  formula_list = NULL,
  log_means,
  log_sds,
  unassigned_name = "__no_feature",
  seqdepth = nfeatures * 2500,
  dispersion = 100,
  lfn_sd = 0.2
)
}
\arguments{
\item{nfeatures}{Number of "features" to simulate data for. A "feature" in this case
may contain a number of "sub-features". For example, you may want to simulate pre-RNA
and mature RNA for a set of "genes", in which case the number of features is the number
of genes.}

\item{graph}{An adjacency matrix describing the reaction diagram graph relating
the various RNA species to one another.}

\item{metadf}{Data frame with two required columns (\code{sample} and \code{tl}). \code{sample}
represents names given to each simulated sample. \code{tl} represents the label time
for that sample. Additional columns can specify other features of the sample,
like what subcellular compartment the sample is taken from. \strong{NOTE: Not sure I am
actually using these optional columns in any useful capacity anymore}.}

\item{formula_list}{A list of named lists. The names of each sub-list should be
the same as the sample names as they are found in \code{metadf}. Each sub-list should be
a list of formula relating feature names that will show up as columns of the simulated
cB to species modeled in your \code{graph}. This only needs to be specified if you want
to simulate the scenario where some of the measured species are a sum of modeled species.}

\item{log_means}{Vector of log-Normal logmeans from which the distribution of
feature-specific parameters will be drawn from. Length of vector should be the same
as max(entries in \code{graph}), i.e., the number of parameters in your specified model.}

\item{log_sds}{Vector of log-Normal logsds from which the distribution of
feature-specific parameters will be drawn from.}

\item{unassigned_name}{String to give to reads not assigned to a given feature.}

\item{seqdepth}{Total number of reads in each sample.}

\item{dispersion}{Negative binomial \code{size} parameter to use for simulating read counts}

\item{lfn_sd}{Logit(fn) replicate variability.}
}
\description{
\code{SimulateDynamics()} simulates any specified dynamical system of interconverting
RNA species. Its required input is similar to that of \code{EstimateDynamics()}, i.e.,
an adjacency matrix describing the set of species and how they are related to
one another and a list of formula relating actually assayed species to the
modeled species. Currently, \code{SimulateDynamics()} implements a naive heteroskedastic
replicate variability simulation and is not designed to simulate multiple experimental
conditions.
}
