% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{SimulateMultiCondition}
\alias{SimulateMultiCondition}
\title{Simulate NR-seq data for multiple replicates of multiple biological conditions}
\usage{
SimulateMultiCondition(
  nfeatures,
  metadf,
  param_details,
  mean_formula,
  seqdepth = 1e+07,
  label_time = 2,
  pnew = 0.05,
  pold = 0.001,
  readlength = 200,
  Ucont = 0.25,
  feature_prefix = "Gene",
  dispslope = 5,
  dispint = 0.01,
  logkdegsdtrend_slope = -0.3,
  logkdegsdtrend_intercept = -1.5,
  logksynsdtrend_slope = -0.3,
  logksynsdtrend_intercept = -1.5
)
}
\arguments{
\item{nfeatures}{Number of "features" (e.g., genes) to simulate data for}

\item{metadf}{A data frame with the following columns:
\itemize{
\item sample: Names given to samples to simulate.
\item \if{html}{\out{<details>}}: Any number of columns with any names (not taken by other metadf columns)
storing factors by which the samples can be stratified. These can be referenced
in \code{mean_formula}, described below.
}
These parameters (described more below) can also be included in metadf to specify sample-specific simulation
parameter:
\itemize{
\item seqdepth
\item label_time
\item pnew
\item pold
\item readlength
\item Ucont
}}

\item{seqdepth}{Only relevant if \code{read_vect} is not provided; in that case, this is
the total number of reads to simulate.}

\item{label_time}{Length of s^{4}U feed to simulate.}

\item{pnew}{Probability that a T is mutated to a C if a read is new.}

\item{pold}{Probability that a T is mutated to a C if a read is old.}

\item{readlength}{Length of simulated reads. In this simple simulation, all reads
are simulated as being exactly this length.}

\item{Ucont}{Probability that a nucleotide in a simulated read is a U.}
}
\description{
\code{SimulateMultiCondition} is a highly flexibly simulator that combines linear modeling
of log(kdeg)'s with \code{SimulateOneRep} to simulate an NR-seq dataset. The linear model
allows you to simulate multiple distinct treatments, batch effects, interaction effects,
etc.
}
