% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations.R
\name{SimulateMultiLabel}
\alias{SimulateMultiLabel}
\title{Simulate one replicate of multi-label NR-seq data}
\usage{
SimulateMultiLabel(
  nfeatures,
  populations = c("TC"),
  fraction_design = create_fraction_design(populations),
  fractions_matrix = NULL,
  read_vect = NULL,
  sample_name = "sampleA",
  feature_prefix = "Gene",
  kdeg_vect = NULL,
  ksyn_vect = NULL,
  logkdeg_mean = -1.9,
  logkdeg_sd = 0.7,
  logksyn_mean = 2.3,
  logksyn_sd = 0.7,
  phighs = setNames(rep(0.05, times = length(populations)), populations),
  plows = setNames(rep(0.002, times = length(populations)), populations),
  seqdepth = nfeatures * 2500,
  readlength = 200,
  alpha_min = 3,
  alpha_max = 6,
  Ucont = 0.25,
  Acont = 0.25,
  Gcont = 0.25,
  Ccont = 0.25
)
}
\description{
Generalizes SimulateOneRep() to simulate any combination of mutation types.
Currently, no kinetic model is used to relate certain parameters to the
fractions of reads belonging to each simulated mutational population. Instead
these fractions are drawn from a Dirichlet distribution with gene-specific
parameters.
}
