% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_fraction_design}
\alias{standard_fraction_design}
\title{Standard \code{fraction_design} table for \code{EstimateFractions}}
\format{
\subsection{\code{standard_fraction_design}}{

A tibble with 2 rows and 2 columns:
\describe{
\item{TC}{Boolean denoting if population represented by that row has high T-to-C mutational content}
\item{present}{Boolean denoting if population represented by that row is expected to be present in this dataset}
}
}
}
\usage{
standard_fraction_design
}
\description{
An example \code{fraction_design} table for a standard NR-seq experiment with
s^4U labeling. This table tells \code{EstimateFractions} that there are two
populations of reads, one with high T-to-C mutation content and one
with low T-to-C mutation content
}
\keyword{datasets}
