% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRArrowData.R
\name{validate_EZbakRArrowData}
\alias{validate_EZbakRArrowData}
\title{\code{EZbakRArrowData} EZbakRArrowData validator}
\usage{
validate_EZbakRArrowData(obj)
}
\arguments{
\item{obj}{An object of class \code{EZbakRArrowData}}
}
\description{
\code{validate_EZbakRArrowData} ensures that input for \code{EZbakRArrowData} object construction
is valid.
}
