% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRFractions.R
\name{validate_EZbakRFractions}
\alias{validate_EZbakRFractions}
\title{\code{EZbakRFractions} object validator}
\usage{
validate_EZbakRFractions(obj)
}
\arguments{
\item{obj}{An object of class \code{EZbakRFractions}}
}
\description{
\code{validate_EZbakRData} ensures that input for \code{EZbakRFractions} construction
is valid.
}
