% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRKinetics.R
\name{validate_EZbakRKinetics}
\alias{validate_EZbakRKinetics}
\title{\code{EZbakRKinetics} object validator}
\usage{
validate_EZbakRKinetics(obj, features)
}
\arguments{
\item{obj}{An object of class \code{EZbakRKinetics}}

\item{features}{Features tracked in \code{kinetics} data frame. Needs to be specified
explicitly as it cannot be automatically inferred.}
}
\description{
\code{validate_EZbakRData} ensures that input for \code{EZbakRKinetics} construction
is valid.
}
