% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{CompareParameters}
\alias{CompareParameters}
\title{Get contrasts of estimated parameters}
\usage{
CompareParameters(
  obj,
  condition,
  reference,
  experimental,
  features = NULL,
  exactMatch = TRUE,
  repeatID = NULL,
  mean_vars = NULL,
  sd_vars = NULL,
  overwrite = TRUE,
  parameter = "log_kdeg"
)
}
\arguments{
\item{obj}{An \code{EZbakRFit} object, which is an \code{EZbakRFractions} object on
which \code{AverageAndRegularize()} has been run.}

\item{condition}{Name of factor from \code{metadf} whose parameter estimates at
different factor values you would like to compare.}

\item{reference}{Name of reference \code{condition} factor level value. Difference
will be calculated as \code{experimental} - \code{reference}.}

\item{experimental}{Name of \code{condition} factor level value to compare to reference.
Difference will be calculated as \code{experimental} - \code{reference}.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{exactMatch}{If TRUE, then \code{features} and \code{populations} have to exactly match
those for a given fractions table for that table to be used. Means that you can't
specify a subset of features or populations by default, since this is TRUE
by default.}

\item{repeatID}{If multiple \code{averages} tables exist with the same metadata,
then this is the numerical index by which they are distinguished.}

\item{mean_vars}{Sample features from metadf that were used in formula for
parameter average linear model.}

\item{sd_vars}{Sample features from metadf that were used in formula for
parameter standard deviation linear model.}

\item{overwrite}{If TRUE, then identical output will be overwritten if it exists.}

\item{parameter}{Parameter to average across replicates of a given condition.}
}
\description{
Get contrasts of estimated parameters
}
