% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{CompareParameters}
\alias{CompareParameters}
\title{Get contrasts of estimated parameters}
\usage{
CompareParameters(
  obj,
  design_factor,
  reference,
  experimental,
  param_name,
  reference_levels = NULL,
  experimental_levels = NULL,
  param_function,
  condition = NULL,
  features = NULL,
  exactMatch = TRUE,
  repeatID = NULL,
  formula_mean = NULL,
  sd_grouping_factors = NULL,
  fit_params = NULL,
  overwrite = TRUE,
  parameter = "log_kdeg",
  type = "averages"
)
}
\arguments{
\item{obj}{An \code{EZbakRFit} object, which is an \code{EZbakRFractions} object on
which \code{AverageAndRegularize()} has been run.}

\item{design_factor}{Name of factor from \code{metadf} whose parameter estimates at
different factor values you would like to compare. If you specify this, you need
to also specify \code{reference} and \code{experimental}.}

\item{reference}{Name of reference \code{condition} factor level value. Difference
will be calculated as \code{experimental} - \code{reference}.}

\item{experimental}{Name of \code{condition} factor level value to compare to reference.
Difference will be calculated as \code{experimental} - \code{reference}.}

\item{param_name}{If you want to assess the significance of a single parameter,
rather than the comparison of two parameters, specify that one parameter's name
here.}

\item{reference_levels}{If type == "dynamics", then this should specify the levels
of the \code{design_factor}(s) reference group. For example, if you have a single
\code{design_factor} called "genotype" and you would like to compare "WT" (reference)
to "KO" (experimental), then \code{reference_levels} should be "WT". If you have
multiple \code{design_factor}'s, then \code{reference_levels} must be a named character vector with one element
per \code{design_factor}, with elements named the corresponding \code{design_factor}. For
example, if \code{design_factor} is c("genotype", "treatment"), and you would like
to compare genotype = "WT" and treatment = "untreated" (reference) to genotype = "KO" and
treatment = "treated", then \code{reference_levels} would need to be a vector with
one element named "genotype", equal to "WT" and one element named "treatment"
equal to "untreated" (this example could be created with, \code{c(genotype = "WT", treatment = "untreated")}).}

\item{experimental_levels}{Same idea as \code{reference_levels}.
If type == "dynamics", then this should specify the levels
of the \code{design_factor}(s) experimental group. For example, if you have a single
\code{design_factor} called "genotype" and you would like to compare "WT" (reference)
to "KO" (experimental), then \code{experimental_levels} should be "KO". If you have
multiple \code{design_factor}'s, then \code{experimental_levels} must be a named character vector with one element
per \code{design_factor}, with elements named the corresponding \code{design_factor}. For
example, if \code{design_factor} is c("genotype", "treatment"), and you would like
to compare genotype = "WT" and treatment = "untreated" (reference) to genotype = "KO" and
treatment = "treated", then \code{experimental_levels} would need to be a vector with
one element named "genotype", equal to "KO" and one element named "treatment"
equal to "treated" (this example could be created with, \code{c(genotype = "KO", treatment = "treated")}).}

\item{param_function}{NOT YET IMPLEMENTED. Will allow you to specify more complicated
functions of parameters when hypotheses you need to test are combinations of parameters
rather than individual parameters or simple differences in two parameters.}

\item{condition}{Same as \code{design_factor}, will be deprecated in favor of the
former in later release.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{exactMatch}{If TRUE, then \code{features} and \code{populations} have to exactly match
those for a given fractions table for that table to be used. Means that you can't
specify a subset of features or populations by default, since this is TRUE
by default.}

\item{repeatID}{If multiple \code{averages} tables exist with the same metadata,
then this is the numerical index by which they are distinguished.}

\item{formula_mean}{An R formula object specifying how the \code{parameter} of interest
depends on the sample characteristics for the averages object you want to use.}

\item{sd_grouping_factors}{Metadf columns should data was grouped by when estimating
standard deviations across replicates for the averages object you want to use.}

\item{fit_params}{Character vector of parameter names in the averages object
you would like to use.}

\item{overwrite}{If TRUE, then identical output will be overwritten if it exists.}

\item{parameter}{Parameter to average across replicates of a given condition.}

\item{type}{Type of table to use. Can either be "averages" or "dynamics".}
}
\description{
Get contrasts of estimated parameters
}
