% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{CompareParameters}
\alias{CompareParameters}
\title{Get contrasts of estimated parameters}
\usage{
CompareParameters(
  obj,
  features = "all",
  parameter = "log_kdeg",
  condition,
  reference,
  experimental
)
}
\description{
Get contrasts of estimated parameters
}
