% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Comparisons.R
\name{CompareParameters}
\alias{CompareParameters}
\title{Get contrasts of estimated parameters}
\usage{
CompareParameters(
  obj,
  condition,
  reference,
  experimental,
  features = NULL,
  overwrite = TRUE,
  parameter = "log_kdeg"
)
}
\arguments{
\item{obj}{An \code{EZbakRFit} object, which is an \code{EZbakRFractions} object on
which \code{AverageAndRegularize()} has been run.}

\item{condition}{Name of factor from \code{metadf} whose parameter estimates at
different factor values you would like to compare.}

\item{reference}{Name of reference \code{condition} factor level value. Difference
will be calculated as \code{experimental} - \code{reference}.}

\item{experimental}{Name of \code{condition} factor level value to compare to reference.
Difference will be calculated as \code{experimental} - \code{reference}.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{overwrite}{If TRUE, then identical output will be overwritten if it exists.}

\item{parameter}{Parameter to average across replicates of a given condition.}
}
\description{
Get contrasts of estimated parameters
}
