% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dropout.R
\name{CorrectDropout}
\alias{CorrectDropout}
\title{Correct for experimental/bioinformatic dropout of labeled RNA.}
\usage{
CorrectDropout(obj, features = NULL, quant_name = NULL)
}
\arguments{
\item{obj}{An EZbakRFractions object, which is an EZbakRData object on which
you have run \code{EstimateFractions()}.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of \code{NULL}
will expect there to be only one fractions table in the EZbakRFractions object.}

\item{quant_name}{Name of quantification tool appended to table name of interest.
This is only relevant if you are providing isoform-specific estimates, in which
case the isoform quantification tool's name may need to be provided in order
for EZbakR to uniquely identify the table of interest. Even in that case though,
this should only have to be non-null in the case where you have performed isoform-specific
fraction estimation with more than one quantification tool's output.}
}
\description{
Correct for experimental/bioinformatic dropout of labeled RNA.
}
