% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{create_fraction_design}
\alias{create_fraction_design}
\title{Generate a \code{fraction_design} table for \code{EstimateFractions}}
\usage{
create_fraction_design(mutrate_populations)
}
\arguments{
\item{mutrate_populations}{Character vector of the set of mutational populations
present in your data.}
}
\value{
A \code{fraction_design} table that assumes that every possible combination of
mutational populations listed in \code{mutrate_populations} are present in your data.
The \code{present} column can be modified if this assumption is incorrect
}
\description{
Generate a \code{fraction_design} table for \code{EstimateFractions}
}
