% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateFractions.EZbakRArrowData}
\alias{EstimateFractions.EZbakRArrowData}
\title{Estimate fractions of each RNA population with standard EZbakRData object}
\usage{
\method{EstimateFractions}{EZbakRArrowData}(
  obj,
  features = "all",
  mutrate_populations = "all",
  fraction_design = NULL,
  Poisson = TRUE,
  strategy = c("standard"),
  column_to_reorder = NULL
)
}
\description{
Estimate fractions of each RNA population with standard EZbakRData object
}
