% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateFractions.EZbakRData}
\alias{EstimateFractions.EZbakRData}
\title{Estimate fractions of each RNA population with standard EZbakRData object}
\usage{
\method{EstimateFractions}{EZbakRData}(
  obj,
  features = "all",
  mutrate_populations = "all",
  fraction_design = NULL,
  Poisson = TRUE,
  strategy = c("standard", "hierarchical"),
  split_multi_features = FALSE,
  multi_feature_cols = NULL,
  multi_feature_sep = "+",
  pnew_prior_mean = -2.94,
  pnew_prior_sd = 0.3,
  pold_prior_mean = -6.5,
  pold_prior_sd = 0.5,
  hier_readcutoff = 300,
  init_pnew_prior_sd = 0.8,
  character_limit = 20,
  overwrite = TRUE
)
}
\description{
Estimate fractions of each RNA population with standard EZbakRData object
}
