% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateFractions.EZbakRData}
\alias{EstimateFractions.EZbakRData}
\title{Estimate fractions of each RNA population with standard EZbakRData object}
\usage{
\method{EstimateFractions}{EZbakRData}(
  obj,
  features = "all",
  mutrate_populations = "all",
  fraction_design = NULL,
  Poisson = TRUE,
  strategy = c("standard")
)
}
\description{
Estimate fractions of each RNA population with standard EZbakRData object
}
