% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fraction_New.R
\name{EstimateFractions}
\alias{EstimateFractions}
\title{Estimate fractions of each RNA population}
\usage{
EstimateFractions(
  obj,
  features = "all",
  populations = "all",
  mutrate_strategy = "standard"
)
}
\arguments{
\item{obj}{EZbakRDataobject}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns.}

\item{populations}{Character vector of the set of mutational populations
that you want to infer the fractions of. For example, say your cB file contains
columns tracking T-to-C and G-to-A}
}
\description{
Estimate fractions of each RNA population
}
