% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateFractions}
\alias{EstimateFractions}
\title{Estimate fractions of each RNA population}
\usage{
EstimateFractions(
  obj,
  features = "all",
  mutrate_populations = "all",
  fraction_design = NULL,
  Poisson = TRUE,
  strategy = c("standard", "hierarchical"),
  filter_cols = "all",
  filter_condition = `&`,
  remove_features = c("NA", "__no_feature"),
  split_multi_features = FALSE,
  multi_feature_cols = NULL,
  multi_feature_sep = "+",
  pnew_prior_mean = -2.94,
  pnew_prior_sd = 0.3,
  pold_prior_mean = -6.5,
  pold_prior_sd = 0.5,
  hier_readcutoff = 300,
  init_pnew_prior_sd = 0.8,
  pnew_prior_sd_min = 0.01,
  pold_est = NULL,
  character_limit = 20,
  overwrite = TRUE
)
}
\arguments{
\item{obj}{\code{EZbakRData} or \code{EZbakRArrowData} object}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB file.}

\item{mutrate_populations}{Character vector of the set of mutational populations
that you want to infer the rates of mutations for. By default, all mutation rates
are estimated for all populations present in cB.}

\item{fraction_design}{"Design matrix" specifying which RNA populations exist
in your samples. By default, this will be created automatically and will assume
that all combinations of the \code{mutrate_populations} you have requested to analyze are
present in your data. If this is not the case for your data, then you will have
to create one manually.

If you call the function \code{create_fraction_design(...)}, providing a vector
of mutational population names as input, it will create a \code{fraction_design} table for
you, with the assumption that every single possible combination of mutational populations
is present in your data. You can then edit the \code{present} column as necessary to
get an appropriate \code{fraction_design} for your use case. See below for details on
the required contents of \code{fraction_design} and its interpretation.

\code{fraction_design} must have one column per element of \code{mutrate_populations},
with these columns sharing the name of the \code{mutrate_populations}. It must also have
one additional column named \code{present}. All elements of fraction_design should be
booleans (\code{TRUE} or \code{FALSE}). It should include all possible combinations of \code{TRUE}
and \code{FALSE} for the \code{mutrate_populations} columns. A \code{TRUE} in one of these columns
represents a population of RNA that is expected to have above background mutation
rates of that type. \code{present} will denote whether or not that population of RNA
is expected to exist in your data.

For example, assume you are doing a typical TimeLapse-seq/SLAM-seq/TUC-seq/etc. experiment
where you have fed cells with s^4U and recoded any incorporated s^4U to a nucleotide
that reverse transcriptase will read as a cytosine. That means that \code{mutrate_populations}
will be "TC", since you want to estimate the fraction of RNA that was s^4U labeled, i.e.,
the fraction with high T-to-C mutation content. \code{fraction_design} will thus have two columns:
\code{TC} and \code{present}. It will also have two rows. One of these rows must have a value of
\code{TRUE} for \code{TC}, and the other must have a value of \code{FALSE}. The row with a value of \code{TRUE}
for \code{TC} represents the population of reads with high T-to-C mutation content,
i.e., the reads from RNA that were synthesized while s^4U was present. The row
with a value of \code{FALSE} for \code{TC} reprsents the population of reads with low T-to-C mutation
content, i.e., the reads from RNA that existed prior to s^4U labeling. Both of these
populations exist in your data, so the value of the \code{present} column should be \code{TRUE} for
both of these. See the lazily loaded \code{standard_fraction_design} object for an example
of what this tibble could look like. ("lazily loaded \code{standard_fraction_design} object"
means that if you run \code{print(standard_fraction_design)} after loading \code{EZbakR} with \code{library(EZbakR)},
then you can see its contents. More specifically, lazily loaded means that this table is not loaded
into memory until you ask for it, via something like a \code{print()} call.)

As another example, consider TILAC, a NR-seq extension developed by the Simon lab. TILAC was originally
described in \href{https://academic.oup.com/nar/article/50/19/e110/6677324}{Courvan et al., 2022}. In this
method, two populations of RNA, one from s^4U fed cells and one from s^6G fed cells, are pooled
and prepped for sequencing together. This allows for internally controlled comparisons of RNA
abundance without spike-ins. s^4U is recoded to a cytosine analog by TimeLapse chemistry
(or similar chemistry) and s^6G is recoded to an adenine analog. Thus, \code{fraction_design} includes
columns called \code{TC} and \code{GA}. A unique aspect of the TILAC \code{fraction_design} table is that
one of the possible populations, \code{TC} and \code{GA} both \code{TRUE}, is denoted as not present (\code{present} = \code{FALSE}).
This is because there is no RNA that was exposed to both s^4U and s^6G, thus a population of reads
with both high T-to-C and G-to-A mutational content should not exist. To see an example
of what a TILAC \code{fraction_design} table could look like, see the lazily loaded
\code{tilac_fraction_design} object.}

\item{Poisson}{If \code{TRUE}, use U-content adjusted Poisson mixture modeling strategy. Often provides
significant speed gain without sacrificing accuracy.}

\item{strategy}{String denoting which new read mutation rate estimation strategy to use.
Options include:
\itemize{
\item standard: Estimate a single new read and old read mutation rate for each
sample. This is done via a binomial mixture model aggregating over
\item hierarchical: Estimate feature-specific new read mutation
rate, regularizing the feature-specific estimate with a sample-wide prior. Currently
only compatible with single mutation type mixture modeling.
}}

\item{split_multi_features}{If a set of reads maps ambiguously to multiple features,
should data for such reads be copied for each feature in the ambiguous set? If this is
\code{TRUE}, then \code{multi_feature_cols} also must be set. Examples where this should be set
to \code{TRUE} includes when analyzing exonic bins (concept defined in original DEXSeq paper),
exon-exon junctions, etc.}

\item{multi_feature_cols}{Character vector of columns that have the potential to
include assignment to multiple features. Only these columns will have their features split
if \code{split_multi_features} is \code{TRUE}.}

\item{multi_feature_sep}{String representing how ambiguous feature assignments are
distinguished in the feature names. For example, the default value of "+" denotes
that if a read maps to multiple features (call them featureA and featureB, for example),
then the feature column will have a value of "featureA+featureB" for that read.}

\item{pnew_prior_mean}{Mean for logit(pnew) prior.}

\item{pnew_prior_sd}{Standard deviation for logit(pnew) prior.}

\item{pold_prior_mean}{Mean for logit(pold) prior.}

\item{pold_prior_sd}{Standard deviation for logit(pold) prior.}

\item{hier_readcutoff}{If \code{strategy} == \code{hierarchical}, only features with this many reads
are used to infer the distribution of feature-specific labeled read mutation rates.}

\item{init_pnew_prior_sd}{If \code{strategy} == \code{hierarchical}, this is the initial logit(pnew)
prior standard deviation to regularize feature-specific labeled read mutation rate estimates.}

\item{character_limit}{Maximum number of characters for naming out fractions output. EZbakR
will try to name this as a "_" separated character vector of all of the features analyzed.
If this name is greater than \code{character_limit}, then it will default to "fraction#", where
"#" represents a simple numerical ID for the table.}

\item{overwrite}{If TRUE and a fractions estimate output already exists that
would possess the same metadata (features analyzed, populations analyzed,
and fraction_design), then it will get overwritten with the new output. Else,
it will be saved as a separate output with the same name + "_#" where "#" is a
numerical ID to distinguish the similar outputs.}
}
\description{
Estimate fractions of each RNA population
}
