% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isoforms.R
\name{EstimateIsoformFractions}
\alias{EstimateIsoformFractions}
\title{Estimate isoform-specific fraction news (or more generally "fractions").}
\usage{
EstimateIsoformFractions(
  obj,
  quantification,
  quant_tool = c("none", "salmon", "sailfish", "alevin", "piscem", "kallisto", "rsem",
    "stringtie"),
  tx2gene = NULL,
  cores = 1
)
}
\arguments{
\item{obj}{An \code{EZbakRData} object}

\item{quantification}{Transcript isoform quantification information. This
can be provided in multiple formats:
\itemize{
\item A data frame with the columns \code{sample}, \code{TPM}, \code{transcript_id}, and \code{gene_id}.
The \code{sample} column should contain the names of all samples present in the cB in
the \code{EZbakRData} object for which fractions have been estimated. \code{TPM} represents
the transcripts per million estimated by the quantification tool of your choice.
\item A data frame with two columns, one named \code{sample} and another named \code{path}.
The \code{sample} column should contain the names of all samples present in the cB in
the \code{EZbakRData} object for which fractions have been estimated. The \code{path}
column should be a path to a transcript isoform quantification file. **If this
form of \code{quantification} is provided, then \code{quant_tool} must be specified so
that the quantification files can be properly interpreted
}}
}
\description{
Combines the output of \code{EstimateFractions} with transcript isoform
quantification performed by an outside tool (e.g., RSEM, kallisto, salmon, etc.)
to infer transcript isoform-specific fraction news (or more generally fraction
of reads coming from a particular mutation population).
}
