% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Isoforms.R
\name{EstimateIsoformFractions}
\alias{EstimateIsoformFractions}
\title{Estimate isoform-specific fraction news (or more generally "fractions").}
\usage{
EstimateIsoformFractions(
  obj,
  features = NULL,
  populations = NULL,
  fraction_design = NULL,
  fraction_name = NULL,
  quant_name = NULL,
  gene_to_transcript = NULL,
  overwrite = TRUE,
  TPM_min = 1,
  count_min = 10
)
}
\arguments{
\item{obj}{An \code{EZbakRData} object}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{populations}{Mutational populations that were analyzed to generate the
fractions table to use. For example, this would be "TC" for a standard
s4U-based nucleotide recoding experiment.}

\item{fraction_design}{"Design matrix" specifying which RNA populations exist
in your samples. By default, this will be created automatically and will assume
that all combinations of the \code{mutrate_populations} you have requested to analyze are
present in your data. If this is not the case for your data, then you will have
to create one manually. See docs for \code{EstimateFractions} (run ?EstimateFractions()) for more details.}

\item{fraction_name}{Name of fraction estimate table to use. Should be stored in the
\code{obj$fractions} list under this name. Can also rely on specifying \code{features} and/or \code{populations}
and having \code{EZget()} find it.}

\item{quant_name}{Name of transcript isoform quantification table to use. Should be stored
in the obj$readcounts list under this name. Use \code{ImportIsoformQuant()} to create
this table. If \code{quant_name} is \code{NULL}, it will search for tables containing the string
"isoform_quant" in their name, as that is the naming convention used by \code{ImportIsoformQuant()}.
If more than one such table exists, an error will be thrown and you will have to specify
the exact name in \code{quant_name}.}

\item{gene_to_transcript}{Table with columns \code{transcript_id} and all feature related
columns that appear in the relevant fractions table. This is only relevant as a hack to
to deal with the case where STAR includes in its transcriptome alignment transcripts
on the opposite strand from where the RNA actually originated. This table will be used
to filter out such transcript-feature combinations that should not exist.}

\item{overwrite}{If TRUE and a fractions estimate output already exists that
would possess the same metadata (features analyzed, populations analyzed,
and fraction_design), then it will get overwritten with the new output. Else,
it will be saved as a separate output with the same name + "_#" where "#" is a
numerical ID to distinguish the similar outputs.}

\item{TPM_min}{Minimum TPM for a transcript to be kept in analysis.}

\item{count_min}{Minimum expected_count for a transcript to be kept in analysis.}
}
\description{
Combines the output of \code{EstimateFractions} with transcript isoform
quantification performed by an outside tool (e.g., RSEM, kallisto, salmon, etc.)
to infer transcript isoform-specific fraction news (or more generally fraction
of reads coming from a particular mutation population).
}
