% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinetics.R
\name{EstimateKinetics}
\alias{EstimateKinetics}
\title{Generic function for estimating kinetic parameters.}
\usage{
EstimateKinetics(
  obj,
  strategy = c("standard", "tilac", "NSS", "pulse-chase"),
  features = NULL,
  populations = NULL,
  fraction_design = NULL,
  grouping_factors = NULL,
  character_limit = 20,
  overwrite = TRUE
)
}
\arguments{
\item{obj}{An \code{EZbakRFractions} object, which is an \code{EZbakRData} object on
which \code{EstimateFractions()} has been run.}

\item{strategy}{Kinetic parameter estimation strategy.
Options include:
\itemize{
\item standard: Estimate a single new read and old read mutation rate for each
sample. This is done via a binomial mixture model aggregating over
\item pulse-chase (NOT YET IMPLEMENTED): Estimate kdeg for a pulse-chase experiment. A kdeg will be estimated
for each time point at which label was present. This includes any pulse-only samples,
as well as all samples including a chase after the pulse.
\item custom (NOT YET IMPLEMENTED): Provide a custom function that takes
fraction estimates as input and produces as output kinetic parameter estimates.
}}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{quant_name}{Name of quantification tool appended to table name of interest.
This is only relevant if you are providing isoform-specific estimates, in which
case the isoform quantification tool's name may need to be provided in order
for EZbakR to uniquely identify the table of interest. Even in that case though,
this should only have to be non-null in the case where you have performed isoform-specific
fraction estimation with more than one quantification tool's output.}
}
\value{
\code{EZbakRKinetics} object.
}
\description{
Generic function for estimating kinetic parameters.
}
