% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kinetics.R
\name{EstimateKinetics}
\alias{EstimateKinetics}
\title{Generic function for estimating kinetic parameters.}
\usage{
EstimateKinetics(obj, features = NULL, strategy = c("standard", "tilac"))
}
\arguments{
\item{obj}{An \code{EZbakRFractions} object, which is an \code{EZbakRData} object on
which \code{EstimateFractions()} has been run.}

\item{features}{Character vector of the set of features you want to stratify
reads by and estimate proportions of each RNA population. The default of "all"
will use all feature columns in the \code{obj}'s cB.}

\item{strategy}{Kinetic parameter estimation strategy.
Options include:
\itemize{
\item standard: Estimate a single new read and old read mutation rate for each
sample. This is done via a binomial mixture model aggregating over
\item pulse-chase (NOT YET IMPLEMENTED): Estimate kdeg for a pulse-chase experiment. A kdeg will be estimated
for each time point at which label was present. This includes any pulse-only samples,
as well as all samples including a chase after the pulse.
\item custom (NOT YET IMPLEMENTED): Provide a custom function that takes
fraction estimates as input and produces as output kinetic parameter estimates.
}}
}
\value{
\code{EZbakRKinetics} object
}
\description{
Generic function for estimating kinetic parameters.
}
