% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fractions.R
\name{EstimateMutRates}
\alias{EstimateMutRates}
\title{Estimate mutation rates}
\usage{
EstimateMutRates(
  obj,
  populations = "all",
  strategy = "standard",
  pnew_prior_mean = -2.94,
  pnew_prior_sd = 0.3,
  pold_prior_mean = -6.5,
  pold_prior_sd = 0.5,
  pold_est = NULL
)
}
\arguments{
\item{obj}{EZbakRDataobject}

\item{populations}{Character vector of the set of mutational populations
that you want to infer the fractions of. For example, say your cB file contains
columns tracking T-to-C and G-to-A}

\item{strategy}{String denoting which new read mutation rate estimation strategy to use.
Options include:
\itemize{
\item standard: Estimate a single new read and old read mutation rate for each
sample. This is done via a binomial mixture model fit to all reads in a given sample.
\item hierarchical (NOT YET IMPLEMENTED): Estimate feature-specific mutation
rate with standard, regularizing the feature-specific
estimate with a sample-wide prior.
\item smalec (NOT YET IMPLEMENTED): Estimate two old read mutation rates, as was done in
Smalec et al., 2023. Idea is that alignment artifacts can give rise to a
high mutation rate old read population that should be accounted for
to avoid overestimating the fraction of new reads
}}

\item{pnew_prior_mean}{logit-Normal mean for logit(pnew) prior.}

\item{pnew_prior_sd}{logit-Normal sd for logit(pnew) prior.}

\item{pold_prior_mean}{logit-Normal mean for logit(pold) prior.}

\item{pold_prior_sd}{logit-Normal sd for logit(pold) prior.}
}
\description{
Estimate mutation rates
}
