% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{EZget}
\alias{EZget}
\title{Easily get EZbakR table of estimates of interest}
\usage{
EZget(
  obj,
  type = c("fractions", "kinetics", "averages", "comparisons"),
  features = NULL,
  populations = NULL,
  fraction_design = NULL,
  parameter = NULL,
  returnNameOnly = FALSE,
  exactMatch = FALSE,
  alwaysCheck = FALSE
)
}
\arguments{
\item{obj}{EZbakRData object}

\item{type}{The class of EZbakR outputs would you like to search through.
Equivalent to the name of the list in the EZbakRData object that contains
the tables of interest.}

\item{features}{Features that must be present in the table of interest.
If \code{exactMatch} is TRUE, then these features must also be the only features
present in the table.}

\item{populations}{Only relevant if \code{type} == "fractions". Mutational
populations that must have been analyzed to generate the table of interest.}

\item{fraction_design}{Only relevant if \code{type} == "fractions". Fraction design
table used to generate the table of interest.}

\item{parameter}{Only relevant if \code{type} == "averages" or "comparisons". Which
parameter was being averaged or compared?}

\item{returnNameOnly}{If TRUE, then only the names of tables that passed your
search criteria will be returned. Else, the single table passing your search
criteria will be returned. If there is more than one table that passes your
search criteria and \code{returnNameOnly} == \code{FALSE}, an error will be thrown.}

\item{exactMatch}{If TRUE, then for \code{features} and \code{populations}, which can be vectors,
ensure that provided vectors of features and populations exactly match the relevant metadata
vectors.}

\item{alwaysCheck}{If TRUE, then even if there is only a single table for the \code{type}
of interest, still run all checks against queries.}
}
\description{
Easily get EZbakR table of estimates of interest
}
