% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{EZQC.EZbakRData}
\alias{EZQC.EZbakRData}
\title{Run quality control checks}
\usage{
\method{EZQC}{EZbakRData}(
  obj,
  mutrate_populations = "all",
  features = "all",
  filter_cols = "all",
  filter_condition = `&`,
  remove_features = c("NA", "__no_feature"),
  ...
)
}
\arguments{
\item{obj}{An EZbakRData object}

\item{mutrate_populations}{Same as in \code{EstimateFractions()}. See \code{?EstimateFractions()}
for details.}

\item{features}{Same as in \code{EstimateFractions()}. See \code{?EstimateFractions()}
for details.}

\item{filter_cols}{Same as in \code{EstimateFractions()}. See \code{?EstimateFractions()}
for details.}

\item{filter_condition}{Same as in \code{EstimateFractions()}. See \code{?EstimateFractions()}
for details.}

\item{remove_features}{Same as in \code{EstimateFractions()}. See \code{?EstimateFractions()}
for details.}

\item{...}{Additional arguments. Currently goes unused.}
}
\value{
A list of \code{ggplot2} objects visualizing the various aspects of your data
assessed by \code{EZQC()}.
}
\description{
Run quality control checks
}
