% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC.R
\name{EZQC.EZbakRFractions}
\alias{EZQC.EZbakRFractions}
\title{Run quality control checks}
\usage{
\method{EZQC}{EZbakRFractions}(obj, features = NULL, populations = NULL, fraction_design = NULL, ...)
}
\arguments{
\item{obj}{EZbakRFractions object, which is an EZbakRData object on which
\code{EstimateFractions} has been run.}

\item{features}{Set of features analyzed in the fractions table you are
interested QCing. This gets passed to \code{EZget()} to help find this table.}

\item{populations}{Set of mutation types analyzed in the fractions table
you are interested in QCing. This gets passed to \code{EZget()} to help find this table.}

\item{fraction_design}{The fraction "design matrix" specified to get the
fractions table you are interested in QCing. This gets passed to \code{EZget()} to
help find this table.}

\item{...}{Additional arguments. Currently goes unused.}
}
\description{
Run quality control checks
}
