% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizations.R
\name{EZVolcanoPlot}
\alias{EZVolcanoPlot}
\title{Make a VolcanoPlot from EZbakR comparison}
\usage{
EZVolcanoPlot(
  obj,
  parameter = "log_kdeg",
  condition = NULL,
  reference = NULL,
  experimental = NULL,
  features = NULL,
  plotlog2 = TRUE,
  FDR_cutoff = 0.05,
  difference_cutoff = 0,
  size = NULL
)
}
\arguments{
\item{obj}{An object of class \code{EZbakRCompare}, which is an \code{EZbakRData} object
on which you have run \code{CompareParameters}}

\item{parameter}{Name of parameter whose comparison you want to plot.}

\item{condition}{Name of condition for which you want to plot the comparison of
two of its levels. Condition should be the name of a column that appears in
your metadf table.}

\item{reference}{Name of reference level for comparison.}

\item{experimental}{Name of experimental level for comparison.}

\item{features}{Character vector of feature names for which comparisons were made.}

\item{plotlog2}{If TRUE, assume that log(parameter) difference is passed in and that
you want to plot log2(parameter) difference. TO-DO: probably best to change this
to a more general scale parameter by which the parameter is multiplied. Default
would be log2(exp(1)) to convert log() to log2().}

\item{FDR_cutoff}{False discovery cutoff by which to color points.}

\item{difference_cutoff}{Minimum absolute difference cutoff by which to color points.}

\item{size}{Size of points, passed to \code{geom_point()} size parameter. If not specified,
a point size is automatically chosen.}
}
\description{
Make a VolcanoPlot from EZbakR comparison
}
