% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_normalized_read_counts.EZbakRFractions}
\alias{get_normalized_read_counts.EZbakRFractions}
\title{Get normalized read counts from either an \code{EZbakRFractions} object.}
\usage{
\method{get_normalized_read_counts}{EZbakRFractions}(
  obj,
  features_to_analyze,
  fractions_name = NULL,
  feature_lengths = NULL
)
}
\arguments{
\item{obj}{An \code{EZbakRFractions} object.}

\item{features_to_analyze}{Features in relevant table}

\item{fractions_name}{Name of fractions table to use}

\item{feature_lengths}{Table of effective lengths for each feature combination in your
data. For example, if your analysis includes features named GF and XF, this
should be a data frame with columns GF, XF, and length.}
}
\description{
Uses TMM normalization strategy, similar to that used by DESeq2 and edgeR.
}
