% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRKinetics.R
\name{new_EZbakRKinetics}
\alias{new_EZbakRKinetics}
\title{\code{EZbakRKinetics} object constructor}
\usage{
new_EZbakRKinetics(
  kinetics,
  features,
  metadf,
  name = NULL,
  character_limit = 20
)
}
\arguments{
\item{kinetics}{Data frame containing information about the kinetic parameters
of interest for each set of features tracked.}

\item{features}{Features tracked in \code{kinetics} data frame. Needs to be specified
explicitly as it cannot be automatically inferred.}

\item{metadf}{Data frame describing each of the samples included
in \code{kinetics}}
}
\description{
\code{new_EZbakRData} efficiently creates an object of class \code{EZbakRData}.
It does not perform any rigorous checks of the legitimacy of this object.
}
