% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{print.EZbakRArrowData}
\alias{print.EZbakRArrowData}
\title{Print method for \code{EZbakRArrowData} objects}
\usage{
\method{print}{EZbakRArrowData}(x, max_name_chars = 60, ...)
}
\arguments{
\item{x}{An \code{EZbakRArrowData} object.}

\item{max_name_chars}{Maximum number of characters to print on each line}

\item{...}{Ignored}
}
\description{
Print method for \code{EZbakRArrowData} objects
}
