% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EZbakRArrowData.R
\name{validate_EZbakRArrowData}
\alias{validate_EZbakRArrowData}
\title{\code{EZbakRDataobject} validator}
\usage{
validate_EZbakRArrowData(obj)
}
\arguments{
\item{obj}{An object of class \code{EZbakRDataobject}}
}
\description{
\code{validate_EZbakRData} ensures that input for \code{EZbakRDataobject} construction
is valid.
}
